backend/requirements.txt
fastapi>=0.115.6
openai-agents>=0.6.2
python-dotenv>=1.0.1
uvicorn[standard]>=0.32.1